Extract top-level nlp package
