+++ src/main/scala/com/microsoft/partnercatalyst/fortis/spark/transforms/locations/LocationsExtractor.scala
@@ -2,7 +2,6 @@ package com.microsoft.partnercatalyst.fortis.spark.transforms.locations
 
 import com.microsoft.partnercatalyst.fortis.spark.transforms.locations.client.FeatureServiceClient
 import com.microsoft.partnercatalyst.fortis.spark.transforms.Location
-import com.microsoft.partnercatalyst.fortis.spark.transforms.locations.nlp.PlaceRecognizer
 
 import scala.collection.mutable
